+++ CHANGELOG.md
@@ -8,6 +8,13 @@ FlatLaf Change Log
 - TabbedPane: In scroll-tab-layout, the cropped line is now hidden. (issue #40)
 - Tree: UI default value `Tree.textBackground` now has a valid color and is no
   longer `null`.
+- Button and ToggleButton: Support per component minimum height (set client
+  property `JComponent.minimumHeight` to an integer). (issue #44)
+- Button and ToggleButton: Do not apply minimum width if button border was
+  changed (is no longer an instance of `FlatButtonBorder`).
+- ToggleButton: No longer use focus width for underline style toggle buttons to
+  compute component size, which reduces/fixes component size in "Flat IntelliJ"
+  and "Flat Darcula" themes.
 
 
 ## 0.24